+++ monday_ml_tutorial/.gitignore
@@ -4,3 +4,4 @@
 
 deepmerge/models
 deepmerge/prepared_data
+wandb
