+++ monday_ml_tutorial/dvc.yaml
@@ -1,0 +1,21 @@
+stages:
+  download_data:
+    cmd: python deepmerge/download_data.py deepmerge/latest_data.fits
+    deps:
+    - deepmerge/download_data.py
+    outs:
+    - deepmerge/latest_data.fits
+  prepare_data:
+    cmd: python deepmerge/prepare_data.py
+    deps:
+    - deepmerge/latest_data.fits
+    - deepmerge/prepare_data.py
+    outs:
+    - deepmerge/prepared_data
+  train_cnn:
+    cmd: python deepmerge/train_cnn.py
+    deps:
+    - deepmerge/prepared_data
+    - deepmerge/train_cnn.py
+    outs:
+    - deepmerge/models/latest
